initial app with registries
Actions, stores and widgets have their own registries. They're typed as any,
which means code that reads from the registry will have to typecast. But typing
as anything else is impossible, since the type would have to be a lowest-common
denominator which then *prevents* upcasting.

This is just a first iteration of the app code. Notably it doesn't yet support
lazy-loading and factories.