Register default store
A default store can be provided when creating an app. This commit
ensures it's added to the registry. Use a symbol (exported as
`DEFAULT_STORE`) to register the store, while still requiring new stores
to be registered using string identifiers. This prevents ID collisions.

Update the various registry interfaces to accept symbols when reading
stores. The `stateFrom` option still only takes a string, since the
default store is already used if `stateFrom` is not provided.

Fixes #27.