Allow default store to be assigned once
Rather than requiring defaultStore to be set when creating the app,
allow it to be assigned once.

For some reason TypeScript wouldn't accept the initializer with just
`defaultStore` moved into the AppMixin. It complained about
`registryProvider` not existing on `instance`. So I've moved that into
yet another WeakMap and added a getter for it instead.