Support default action stores
Allow a default action store to be specified. This makes it impossible
to automatically observe the (defined or default) store, so instead the
store is passed to the action factories. It is now their responsibility
to observe the store (if any).

Default action stores work for registered action factories, including
those from action definitions.