Simplify action options and configuration
Pass an options object to action factories. Like the options for widget
factories this object contains a RegistryProvider and the store the
produced action should observe for its state.

Also pass RegistryProvider when configuring actions.

This brings actions more in line with widgets, which are also supposed
to work with the RegistryProvider. The CombinedRegistry approach was
used before the RegistryProvider existed.