Create widgets via registry (#26)
* wip! Add create() to widget registry

create() receives a factory and creation options. It adds the registry
provider to the creation options as well as the default store (if any),
if an ID was specified.

Creates the widget and registers it internally, generating a unique
ID if necessary. Returns a promise for a tuple of the ID and widget.

Fixes #22.

TODO:
* Update README
* Tests

* Comment out Safari for SauceLabs

* SauceLabs no likes edge either

* Creating a widget with app now attempts to add state

When a widget is created and an ID is supplied in the
options but without a `stateFrom` the app will attempt
to `.add()` any state to the default widget store so
that the child widget will have an initiailised state
when it is created.

* Add unit tests for app#createWidget()

* Attempt add on non-default widget stores

* Fix MID for tests/support/createApp