Restore most of the original getWidget implementation
Coverage for the else-branch of
<https://github.com/dojo/app/blob/9dc8c03b1fe63b9054aca017ca47fd4fe3c2dbee/src/createApp.ts#L753:L755>
broke (see
<https://github.com/dojo/app/blob/9dc8c03b1fe63b9054aca017ca47fd4fe3c2dbee/tests/unit/createApp/widgets.ts#L425>)
because the registry was no longer called synchronously (see #55). Of
course the test could have been amended, though the promise interaction
would've been tricky.

Restoring most of the original implementation surfaced another error
that was swallowed by the `.catch()` chaining introduced in #55.
createCustomWidget() would crash if there was no default widget store.
I've added a guard to ensure createCustomWidget() is only called if
there is a default store.