Refactor getWidget() again
Follow-up to #59.

Add a test to verify that errors caused by calling factory are
propagated, without creating a custom widget. This test fails with the
original implementation in #55.

Restore most of the promise chaining from #55, which is more readable
than the nested try/catch statements.

Retain the defaultWidgetStore guard introduced in #59, but acknowledge
that errors thrown by createCustomWidget are masked. In other words,
remove the very subtle control flow that allowed createCustomWidget
errors to bubble up to the calling code. This may hide bugs but should
generally be OK.